+++ train.c
@@ -11,7 +11,7 @@ void read_csv(FILE *f, Dataset *ds_spam, Dataset *ds_ham)
     // Ignore first row
     fgets(buffer, ROW_SIZE, f);
 
-    int i = 0;
+    size_t i = 0;
     while (!feof(f) && i < TS_SIZE) {
         fgets(buffer, ROW_SIZE, f);
 
@@ -35,7 +35,7 @@ void read_csv(FILE *f, Dataset *ds_spam, Dataset *ds_ham)
 void populate_dict(Dict *dict, Dataset *ds, const char *sep, bool spam)
 {
     char *key;
-    for (int i = 0; i < ds->size; i++) {
+    for (size_t i = 0; i < ds->size; i++) {
         key = strtok(ds->rows[i], sep);
         while (key != NULL) {
             for (char *p = key; *p; p++)
@@ -51,7 +51,7 @@ void populate_dict(Dict *dict, Dataset *ds, const char *sep, bool spam)
 void inc_dict(Dict *dict, char *key, bool spam)
 {
     bool found = false;
-    int i;
+    size_t i;
     for (i = 0; i < dict->size; i++) {
         if (strcmp(dict->keys[i], key) == 0) {
             if (spam)
@@ -80,7 +80,7 @@ void get_counts(Dict *dict, char *key, size_t *spam_count, size_t *ham_count)
     *spam_count = 0;
     *ham_count = 0;
     bool found = false;
-    for (int i = 0; i < dict->size; i++) {
+    for (size_t i = 0; i < dict->size; i++) {
         if (strcmp(dict->keys[i], key) == 0) {
             *spam_count = dict->spam_counts[i];
             *ham_count = dict->ham_counts[i];
